+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # A VENIR
 
-
+- [ ] pyweatherfr moscou -w -j50
 
 
 - [ ] emojis compatibles windows et linux